Introducing endpoints parallelization build
With this commit it is introduced a parallelization for the endpoints
regeneration. The daemon will have worker threads as the same number of
CPUs available on the system. Each of those worker threads will
signalize the endpoints turn to build itself.

Thus, the endpoints creation/regeneration will be asynchronous where the
user should verify if the endpoint had a successful regeneration with
`cilium endpoint list`.

Due the parallelization created around the endpoints builder there was a
need to create a mutex on various structures. Those mutex should be used
whenever that same structure's values are read or written.
All private methods of those structures will assume the structure locks
are locked, all public methods will assmue those structures locks are
unlocked, except when stated otherwise such as "MethodLocked", where
the public method needs to be called with the lock locked.

Signed-off-by: André Martins <andre@cilium.io>